+++ src/hooks/__tests__/useObject.test.tsx
@@ -1,95 +1,21 @@
-import React, { FC } from 'react'
+import React from 'react'
 import { render, wait } from '@vtex/test-tools/react'
 
-import { useObject } from '../useObject'
+import { MockObject } from '../../__mocks__/mockObjectComponent'
 import { FormContext } from '../../components'
-import {
-  UseRawInputReturnType,
-  InputReturnTypes,
-  InputTypes,
-  UseRadioReturnType,
-  UseSelectReturnType,
-  UISchemaType,
-  UITypes,
-} from '../types'
-import mockObjectSchema, { toFixed } from '../__mocks__/mockSchema'
+import mockObjectSchema from '../__mocks__/mockSchema'
+import { UISchemaType, UITypes } from '../types'
 
 const mockUISchema: UISchemaType = {
   type: UITypes.default,
   properties: {
+    stringTest: {
+      type: UITypes.radio,
+    },
     numberTest: {
       type: UITypes.select,
     },
   },
-}
-
-const SpecializedObject: FC<{ baseObject: InputReturnTypes }> = props => {
-  switch (props.baseObject.type) {
-    case InputTypes.input: {
-      const inputObject = props.baseObject as UseRawInputReturnType
-      return (
-        <>
-          <label {...inputObject.getLabelProps()}>{inputObject.name}</label>
-          <input {...inputObject.getInputProps()} />
-        </>
-      )
-    }
-    case InputTypes.radio: {
-      const radioObject = props.baseObject as UseRadioReturnType
-      return (
-        <>
-          <label {...radioObject.getLabelProps()}>{radioObject.name}</label>
-          {radioObject.getItems().map((value, index) => {
-            return (
-              <label
-                {...radioObject.getItemLabelProps(index)}
-                key={`${value}${index}`}
-              >
-                {value}
-                <input {...radioObject.getItemInputProps(index)} />
-              </label>
-            )
-          })}
-        </>
-      )
-    }
-    case InputTypes.select: {
-      const selectObject = props.baseObject as UseSelectReturnType
-      return (
-        <>
-          <label {...selectObject.getLabelProps()}>{selectObject.name}</label>
-          <select {...selectObject.getSelectProps()}>
-            {selectObject.getItems().map((value, index) => {
-              return (
-                <option
-                  {...selectObject.getItemOptionProps(index)}
-                  key={`${value}${index}`}
-                >
-                  {value}
-                </option>
-              )
-            })}
-          </select>
-        </>
-      )
-    }
-  }
-  return <></>
-}
-
-const MockObject: FC<{ path: string; UISchema?: UISchemaType }> = props => {
-  const methods = useObject({ path: props.path, UISchema: props.UISchema })
-
-  return (
-    <>
-      {methods.map(obj => (
-        <div key={`${obj.type}${obj.path}`}>
-          <SpecializedObject baseObject={obj} />
-          {obj.getError() && <p>This is an error!</p>}
-        </div>
-      ))}
-    </>
-  )
 }
 
 test('should render all child properties of the schema', () => {