+++ CHANGELOG.md
@@ -7,6 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0-beta.1] - 2020-01-29
+
+### Added
+
+- Added `useCheckbox` hook
+
+### Changed
+
+- Changed the BasicInputReturnType to also return a reference to the validator used
+- **BREAKING**: The `useObject` hook now automatically associates a boolean to a checkbox
+
+### Fixed
+
+- Fixed returned form data not converting string values that represent booleans to actual booleans
+
 ## [0.2.0-beta.0] - 2020-01-23
 
 ### Changed